requirements-gpu.txt
# GPU-Optimized Requirements for RunPod PyTorch 2.4.0
# Compatible avec runpod/pytorch:2.4.0-py3.11-cuda12.4.1-devel-ubuntu22.04

# Base requirements (same as current)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
pydantic==2.5.0
aiofiles==23.2.1

# Fallback CPU whisper (for compatibility)
openai-whisper==20231117

# System monitoring
psutil==5.9.6

# Transcription GPU - Compatible avec PyTorch 2.4.0
faster-whisper>=0.10.0

# PyTorch - Utiliser les versions déjà installées dans RunPod
# torch==2.4.0  # Déjà installé dans l'image RunPod
# torchaudio==2.4.0  # Déjà installé dans l'image RunPod
transformers>=4.36.0

# Diarisation pyannote - Versions compatibles PyTorch 2.4.0
pyannote.audio>=3.1.0,<3.4.0
pyannote.core>=5.0.0
pyannote.database>=5.0.0
pyannote.metrics>=3.2.0

# Audio processing - Enhanced
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.0

# Clustering et ML - Compatible versions
scikit-learn>=1.3.0
numba>=0.58.0
numpy>=1.24.0

# Memory optimization for large models
memory-profiler==0.61.0

# Note: speechbrain et asteroid peuvent causer des conflits
# On les ajoute optionnellement si nécessaire
# speechbrain>=0.5.16
# asteroid>=0.7.0 